Add support for Groth16 shrink-wrapping to the actor system
