+++ risc0/r0vm/Cargo.toml
@@ -17,6 +17,8 @@ bytemuck = "1.12"
 clap = { version = "4.5", features = ["derive", "env"] }
 derive_more = { version = "2.0.1", default-features = false, features = [
   "debug",
+  "from",
+  "try_into",
 ] }
 dirs = "6.0.0"
 futures = "0.3.31"
