+++ risc0/r0vm/src/actors/protocol.rs
@@ -15,39 +15,42 @@
 use std::{ops::Range, sync::Arc};
 
 use clap::ValueEnum;
-use derive_more::Debug;
+use derive_more::{Debug, TryInto};
 use kameo::{actor::ActorRef, Reply};
 use risc0_zkvm::{
-    ProveKeccakRequest, ReceiptClaim, Segment, SegmentReceipt, SuccinctReceipt, UnionClaim, Unknown,
+    ProveKeccakRequest, Receipt, ReceiptClaim, Segment, SegmentReceipt, SuccinctReceipt,
+    UnionClaim, Unknown,
 };
 use serde::{Deserialize, Serialize};
 
 use super::job::JobActor;
 
-pub use risc0_zkvm::rpc::{JobInfo, JobStatus, ProofRequest, ProofResult, Session, TaskError};
+pub use risc0_zkvm::rpc::{
+    JobInfo, JobStatus, ProofRequest, ProofResult, Session, ShrinkWrapKind, ShrinkWrapRequest,
+    ShrinkWrapResult, TaskError,
+};
 
 pub(crate) type JobId = uuid::Uuid;
 pub(crate) type TaskId = u64;
 pub(crate) type WorkerId = uuid::Uuid;
 
 #[derive(Serialize, Deserialize)]
-pub(crate) struct CreateJobRequest {
-    pub request: ProofRequest,
-}
-
-#[derive(Reply, Serialize, Deserialize)]
-pub(crate) struct CreateJobReply {
-    pub job_id: JobId,
-}
-
-#[derive(Serialize, Deserialize)]
 pub(crate) struct JobStatusRequest {
     pub job_id: JobId,
 }
 
-#[derive(Reply, Serialize, Deserialize)]
-pub(crate) struct JobStatusReply {
-    pub info: Option<JobInfo>,
+#[derive(Reply, Serialize, Deserialize, Debug, Clone)]
+pub(crate) struct JobRequestReply {
+    pub job_id: JobId,
+    pub status: JobStatusReply,
+}
+
+#[derive(Reply, Serialize, Deserialize, TryInto, Debug, Clone)]
+pub(crate) enum JobStatusReply {
+    Proof(JobInfo<ProofResult>),
+    ShrinkWrap(JobInfo<ShrinkWrapResult>),
+    #[try_into(ignore)]
+    NotFound,
 }
 
 #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, Serialize, Deserialize, ValueEnum)]
@@ -59,6 +62,7 @@ pub(crate) enum TaskKind {
     Join,
     Union,
     Resolve,
+    ShrinkWrap,
 }
 
 #[derive(Clone, Copy, Hash, PartialEq, Eq, Serialize, Deserialize)]
@@ -82,6 +86,7 @@ pub(crate) enum Task {
     Join(Arc<JoinTask>),
     Union(Arc<UnionTask>),
     Resolve(Arc<ResolveTask>),
+    ShrinkWrap(Arc<ShrinkWrapTask>),
 }
 
 #[derive(Reply, Serialize, Deserialize)]
@@ -128,6 +133,12 @@ pub(crate) struct UnionTask {
 pub(crate) struct ResolveTask {
     pub conditional: Arc<SuccinctReceipt<ReceiptClaim>>,
     pub assumption: Arc<SuccinctReceipt<Unknown>>,
+}
+
+#[derive(Reply, Serialize, Deserialize)]
+pub(crate) struct ShrinkWrapTask {
+    pub kind: ShrinkWrapKind,
+    pub receipt: Arc<Receipt>,
 }
 
 pub mod factory {
@@ -180,6 +191,7 @@ pub mod factory {
         Join(Box<JoinNode>),
         Union(Arc<UnionDone>),
         Resolve(Arc<SuccinctReceipt<ReceiptClaim>>),
+        ShrinkWrap(Arc<Receipt>),
     }
 
     #[derive(Serialize, Deserialize)]
@@ -240,6 +252,7 @@ impl Task {
             Task::Join(_) => TaskKind::Join,
             Task::Union(_) => TaskKind::Union,
             Task::Resolve(_) => TaskKind::Resolve,
+            Task::ShrinkWrap(_) => TaskKind::ShrinkWrap,
         }
     }
 }
